+++ app/main.py
@@ -2,13 +2,10 @@ from fastapi import FastAPI
 from fastapi.middleware.cors import CORSMiddleware
 import uvicorn
 
-from app import db, ml, viz
+from . import db, ml, viz, records
 
 description = """
-This is our FastAPI DS API. to use the current api send a post request to 
-/getdata endpoint with the member_id. a predicted exit destination along with 
-top features will show in a JSON format. Below is an example on how to use it 
-interactively. 
+Edit your app's title and description. See [https://fastapi.tiangolo.com/tutorial/metadata/](https://fastapi.tiangolo.com/tutorial/metadata/)
 
 To use these interactive docs:
 - Click on an endpoint below
@@ -16,22 +13,17 @@ To use these interactive docs:
 - Edit the Request body or any parameters
 - Click the **Execute** button
 - Scroll down to see the Server response Code & Details
-
-To edit description after some endpoints has been added.
-Edit your app's title and description. See [https://fastapi.tiangolo.com/tutorial/metadata/](https://fastapi.tiangolo.com/tutorial/metadata/)
-
 """
 
 app = FastAPI(
-    title='Family Promise of Spokane - Labs 29 ',
+    title='DS API',
     description=description,
     docs_url='/',
 )
-# tags to show on FatsAPI 
-# app.include_router(db.router, tags=['Database'])
+
+app.include_router(records.router, tags=['Database'])
 app.include_router(ml.router, tags=['Machine Learning'])
 app.include_router(viz.router, tags=['Visualization'])
-#app.include_router(viz.router, tags=['Shapley'])
 
 app.add_middleware(
     CORSMiddleware,
